requirements.txt
# local package
-e .

# external requirements (please freeze version to the one you are using)
bandit>=1.7.0
black>=21.10b0
doit>=0.33.1
dvc>=2.8.1
flake8>=3.9.1
loguru>=0.5.3
mkdocs-material>=8.1.9
mkdocs-git-revision-date-plugin>=0.3.1
mlflow>=1.21.0
pylint>=2.11.1
pytest>=6.1.1
pytest-cov>=2.11.1
python-dotenv>=0.19.1
tokenize-rt>=4.2.1
jupyterlab>=3.4.0
jupyterhub>=2.3.0
notebook>=6.4.11 

# Add new requirements below
openslide-python
scikit-image
scikit-learn
numpy
pillow
tensorboard
matplotlib
pandas
tqdm
tensorflow
django
